separated the routes from index.js
