+++ package.json
@@ -9,7 +9,6 @@
     "lint": "ng lint",
     "e2e": "ng e2e",
     "postinstall": "ng build --output-hashing none"
-
   },
   "private": true,
   "dependencies": {
